Four conflicts, all of them two prompt raises landing at once. #840 gave
Echo object emoji while this branch gave both profiles the org source,
and the two are additive rather than competing, so both budget entries
and both raises stand.

The budgets are ratchets rather than measurements, so I re-rendered the
snapshots from the merged sources and set the numbers to what they
actually measure. The delta is 1951 bytes on each profile against main's
own rendered files, identical on both, which is the shared source
showing up as arithmetic.

`docs/sirens-echo-prompt-budget.md` crossed the documentation cap as a
result. That file is an append-only ledger wrapped around a short policy
statement, so it would cross again on the next raise no matter whose it
was. The ledger moves to its own file and the policy doc keeps a pointer.

+++ internal/community/jobhttp.go
@@ -53,7 +53,7 @@ func (a *Agent) handleJobs(writer http.ResponseWriter, request *http.Request) {
 		http.Error(writer, "method not allowed", http.StatusMethodNotAllowed)
 		return
 	}
-	request.Body = http.MaxBytesReader(writer, request.Body, maxHTTPBody)
+	request.Body = http.MaxBytesReader(writer, request.Body, int64(maxHTTPBody))
 	var payload jobSubmitRequest
 	if err := json.NewDecoder(request.Body).Decode(&payload); err != nil {
 		if oversizeBody(err) {
